+++ src/handlers/handle_chr.c
@@ -11,12 +11,14 @@
 /* ************************************************************************** */
 
 #include "ft_printf.h"
+#include "string/ft_str.h"
+#include "char/ft_char.h"
 
-ssize_t		ft_handle_chr(t_params *params, va_list args)
+int		ft_handle_chr(t_params *params, va_list args)
 {
 	char	*pad;
 	char	out;
-	ssize_t	ret;
+	int		ret;
 
 	pad = ft_strnew(1);
 	out = (char)va_arg(args, int);
